Creates a SpaceCollection class to replace HashSet extensions.
